+++ test/classes/test_temporalnetwork.py
@@ -94,6 +94,9 @@ def test_tnet_functions():
     D = tnet.calc_networkmeasure('temporal_degree_centrality')
     assert all(G.sum(axis=-1).sum(axis=-1) == D)
 
+
+
+
 def test_generatenetwork():
     tnet = teneto.TemporalNetwork()
     tnet.generatenetwork('rand_binomial', size=(5,10), prob=1)